requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.3.0
azure-cognitiveservices-speech==1.41.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
eventlet==0.38.2
fastapi==0.115.6
Flask==3.1.0
Flask-SocketIO==5.4.1
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
keyboard==0.13.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.1.0
numpy==2.2.0
opencv-python==4.10.0.84
pillow==11.0.0
propcache==0.2.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytweening==1.2.0
pywin32==308
pywinauto==0.6.8
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
